extract scanner code
